changed vfs2vdc.c

+++ math.h
@@ -1,4 +1,4 @@
-int my_ceil(double x) {
+double my_ceil(double x) {
     int num = (int)x;
     if(x == (double)num)
         return num;
